% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode.R
\name{get_latlon}
\alias{get_latlon}
\title{Get Response from the 'postcodes.io' API and Parse It}
\usage{
get_latlon(postcode)
}
\arguments{
\item{postcode}{Character scalar. A UK postcode.}
}
\value{
If the API call is successfult, a list with two numeric elements:
\code{lat} and \code{lon}.
}
\description{
Get Response from the 'postcodes.io' API and Parse It
}
\examples{
\dontrun{get_latlon("wc2n5du")}
}
