% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weva-package.R
\docType{package}
\name{weva-package}
\alias{weva}
\alias{weva-package}
\title{weva: A Limited and Opinionated Weather-Report CLI}
\description{
Fetch weather information for a UK postcode and print it as a one-line report. Intended primarily for use as a limited and opinionated command-line interface (CLI). Gets weather information from the 'Open-Meteo' API and handles geolocation with the 'postcodes.io' API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/weva}
  \item Report bugs at \url{https://github.com/matt-dray/weva/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray+weva@gmail.com}

}
\keyword{internal}
