% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{get_weather}
\alias{get_weather}
\title{Get Response from the 'Open-Meteo' API and Parse It}
\usage{
get_weather(latlon)
}
\arguments{
\item{latlon}{A list with two numeric elements: 'lat' and 'lon'. Returned by
\link{get_latlon}.}
}
\value{
If the API call is successful, a list with 13 elements, including
\code{current}, \code{hourly} and \code{daily}.
}
\description{
Get Response from the 'Open-Meteo' API and Parse It
}
\examples{
\dontrun{
latlon <- get_latlon("WC2N 5DU")
get_weather(latlon)
}
}
