% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_weather}
\alias{get_weather}
\title{Get Weather Response from 'Open-Meteo'}
\usage{
get_weather(postcode)
}
\arguments{
\item{postcode}{Character scalar. A UK postcode.}
}
\description{
Get Weather Response from 'Open-Meteo'
}
