% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{prepare_report}
\alias{prepare_report}
\title{Prepare Weather Report}
\usage{
prepare_report(response, hours, extremes)
}
\arguments{
\item{response}{A list. Response to a query from the 'Open-Meteo' API,
fetched with \link{get_weather}.}

\item{hours}{Integer scalar. How many hours after now for the 'later' report.}

\item{extremes}{Logical scalar. Show today's min/max temperatures?}
}
\value{
Character scalar.
}
\description{
Prepare data into a micro weather-report string with 'now' and 'later'
segments, and possibly a segment for today's min- and max-temperature.
}
\examples{
\dontrun{get_weather("wc2n5du") |> prepare_report(3L, TRUE)}
}
