% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weva-package.R
\docType{package}
\name{weva-package}
\alias{weva}
\alias{weva-package}
\title{weva: A Tiny Weather-Report CLI}
\description{
A limited and opinionated command-line interface (CLI) to fetch weather information for a UK postcode from the 'Open-Meteo' API and present it as a single-line in the terminal.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/weva}
  \item Report bugs at \url{https://github.com/matt-dray/weva/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray+weva@gmail.com}

}
\keyword{internal}
